+++ migrations/20260206000000_offline_tiles.down.sql
@@ -1,0 +1,2 @@
+DROP TABLE offline_tile;
+DROP INDEX tile_x_y_z_index;